using alpine image , using multi stage build
In this  dockerfile using multi-stage builds, we first use a smaller base image FROM node:18-alpine instead of FROM node:18.15.0@sha256:eee4c74962a855bc1694676cd5069460e545a8580e0e574b2c59eae4f40bd5bd. This reduces the size of the image and improves its security.

In the first stage, we copy the required files and install the dependencies using Yarn. Then, we build the required components of the application 

In the second stage, we copy the required files from the first stage into a new base image, node:18-alpine, and set the working directory to the backend-ws directory.