allow leading zeros (int > decimal edge case)
if a user has an integer value in the send field and then tries to make it into a decimal instead, they might try and put "0." at the start of the number. However, the 0 would not be allowed with this rule.

To get around this they had to just add a . at the start, which would automatically prepend the leading 0. However, some users might not be aware of this, so just making the restrictions looser for now